requirements.txt
Django==1.8.6
MySQL-python==1.2.5
Pillow==3.0.0
Pygments==2.0.2
argparse==1.2.1
blessings==1.6
bpython==0.14.2
curtsies==0.1.19
decorator==4.0.4
djangorestframework==3.3.1
greenlet==0.4.9
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wsgiref==0.1.2


